tiny-poppler: inject DPI metadata do PNG
